Cleaned Up Movement and Created Emitter Prefab
Movement is pretty precise now. Will add head bob later down the line in the polishing phase. Next mechanic is the shooting of the gravity emitter.

Decided to shoot the emitter in a linear path, which will be affected by other gravity fields. Planning on adding trajectory path of emitter.